refactor(observation): decompose runs_service god-file into submodules
runs_service.rs was 1884 lines (god_file). It already contained the
decomposition seams as inline `mod` blocks (run_lookup, artifact_links,
artifact_resolve, persisted_cleanup, runner_downloads) plus shared types,
retain_terminal_runs, and a large test module.

Move each inline module into its own file under runs_service/, keeping
the shared types, retain_terminal_runs, and `mod X; pub use X::*;`
re-exports in mod.rs. The public API is byte-identical (every prior
export is preserved via glob re-export) and no logic changed — this is a
pure file-split.

Largest resulting file is 576 lines (tests); all are well under the
1500-line god_file threshold.

Clears the runs_service.rs god_file finding.
